third_party/blink/web_tests/http/tests/security/script-onerror-crossorigin-same-origin-expected.txt

CONSOLE ERROR: Uncaught SomeError
The test passes if 'window.onerror' is invoked with unsanitized information on a script loaded with a 'crossorigin' attribute, but loads from same-origin and without valid CORS headers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS /SomeError/.test(msg) is true
PASS url is "http://127.0.0.1:8000/security/resources/cors-script.php?fail=true&cors=false"
PASS line is 1
PASS column is 1
PASS window.errorObject is not null
PASS successfullyParsed is true

TEST COMPLETE
